refactor: 레이아웃 관련 컴포넌트들 폴더 구조 변경
