+++ src/components/Layout/Layout.module.scss
@@ -1,0 +1,21 @@
+.header {
+  width: 100%;
+}
+
+.form {
+  padding: 0 1rem;
+  background-color: lighten($bg-color, 5%);
+}
+
+.input_container {
+  display: flex;
+
+  input {
+    flex: 1;
+    height: 3.6rem;
+  }
+}
+
+.header_title {
+  @include blind;
+}
