--- src/components/layout.module.scss
@@ -1,7 +1,0 @@
-.header {
-  width: 100%;
-}
-
-.header_title {
-  @include blind;
-}
